Make only one generate package function
