+++ package_exporter/src/lib.rs
@@ -1,9 +1,7 @@
 mod generate_package;
-mod generate_package_2;
 mod template;
 mod ros2_package_path;
 mod urdf;
 
 pub use generate_package::generate_package;
 pub use template::{PackageContext, Person};
-pub use generate_package_2::generate_package_2;
